requirements.txt
yfinance>=0.2.32
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
